refact(admins): moving it to modules from feature
