--- dashboard/src/features/admins/types/admin-prop.ts
@@ -1,5 +1,0 @@
-import { AdminType } from "@marzneshin/features/admins";
-
-export interface AdminProp {
-    admin: AdminType;
-}
