+++ dashboard/src/modules/admins/components/pills/index.ts
