+++ dashboard/src/modules/admins/components/tables/admins/columns.tsx
@@ -3,7 +3,7 @@ import {
     type AdminType,
     AdminEnabledPill,
     AdminPermissionPill,
-} from "@marzneshin/features/admins";
+} from "@marzneshin/modules/admins";
 import {
     DataTableActionsCell,
     DataTableColumnHeader
