+++ dashboard/src/modules/admins/components/tables/admins/index.tsx
@@ -1,6 +1,6 @@
 import type { FC } from "react";
 import { ColumnDef } from "@tanstack/react-table";
-import { fetchAdmins, AdminType, AdminsQueryFetchKey } from "@marzneshin/features/admins";
+import { fetchAdmins, AdminType, AdminsQueryFetchKey } from "@marzneshin/modules/admins";
 import { columns as columnsFn } from "./columns";
 import { EntityTable } from "@marzneshin/features/entity-table";
 import { useNavigate } from "@tanstack/react-router";
