+++ dashboard/src/modules/admins/contexts/router.ts
@@ -1,5 +1,5 @@
 import { createContext, useContext } from "react";
-import { AdminType } from "@marzneshin/features/admins";
+import { AdminType } from "@marzneshin/modules/admins";
 
 interface RouterAdminContextProps {
     admin: AdminType | null;
