+++ dashboard/src/modules/admins/services/admin.query.ts
