+++ dashboard/src/modules/admins/types/admin-prop.ts
@@ -1,0 +1,5 @@
+import { AdminType } from "@marzneshin/modules/admins";
+
+export interface AdminProp {
+    admin: AdminType;
+}
