+++ dashboard/src/modules/admins/types/schema.ts
