+++ dashboard/src/routes/_dashboard/admins/$adminId/index.tsx
@@ -5,7 +5,7 @@ import {
 import {
     useRouterAdminContext,
     AdminsSettingsDialog,
-} from "@marzneshin/features/admins";
+} from "@marzneshin/modules/admins";
 import { useDialog } from "@marzneshin/hooks";
 
 const AdminOpen = () => {
